Introduce littlefs2-core crate
To make it easier to maintain applications using littlefs2, this patch
introduces a littlefs2-core crate with the basic types and traits used
by littlefs2, but without a specific implementation.

Fixes: https://github.com/trussed-dev/littlefs2/issues/55
