+++ Cargo.toml
@@ -1,27 +1,30 @@
+[workspace]
+members = ["core"]
+
+[workspace.package]
+edition = "2021"
+license = "Apache-2.0 OR MIT"
+repository = "https://github.com/trussed-dev/littlefs2"
+
 [package]
 name = "littlefs2"
 description = "Idiomatic Rust API for littlefs"
 version = "0.4.0"
 authors = ["Nicolas Stalder <n@stalder.io>", "Brandon Edens <brandonedens@gmail.com>", "The Trussed developers"]
-edition = "2021"
-license = "Apache-2.0 OR MIT"
 readme = "README.md"
 categories = ["embedded", "filesystem", "no-std"]
-repository = "https://github.com/trussed-dev/littlefs2"
 documentation = "https://docs.rs/littlefs2"
 
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
 [dependencies]
-bitflags = "1"
 delog = "0.1.0"
 generic-array = "0.14"
 heapless = "0.7"
+littlefs2-core = { version = "0.1", path = "core" }
 littlefs2-sys = "0.2"
-
-[dependencies.serde]
-version = "1"
-default-features = false
-features = ["derive"]
-optional = true
 
 [dev-dependencies]
 ssmarshal = "1"
@@ -31,7 +34,8 @@ serde = { version = "1.0", default-features = false, features = ["derive"] }
 [features]
 default = ["dir-entry-path", "serde"]
 # use experimental closure-based API
-dir-entry-path = []
+dir-entry-path = ["littlefs2-core/dir-entry-path"]
+serde = ["littlefs2-core/serde"]
 # enable assertions in backend C code
 ll-assertions = ["littlefs2-sys/assertions"]
 # enable trace in backend C code
