+++ core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "littlefs2-core"
+version = "0.1.0"
+authors = ["The Trussed developers"]
+description = "Core types for the littlefs2 crate"
+
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+bitflags = "2.6.0"
+generic-array = "0.14"
+heapless = "0.7"
+serde = { version = "1", default-features = false, features = ["derive"], optional = true }
+
+[features]
+dir-entry-path = []
+serde = ["dep:serde"]
