+++ core/src/consts.rs
@@ -1,0 +1,5 @@
+pub const PATH_MAX: usize = 255;
+pub const PATH_MAX_PLUS_ONE: usize = PATH_MAX + 1;
+pub const ATTRBYTES_MAX: u32 = 1_022;
+#[allow(non_camel_case_types)]
+pub type ATTRBYTES_MAX_TYPE = generic_array::typenum::consts::U1022;
