--- src/io/prelude.rs
@@ -1,3 +1,0 @@
-//! Export of the Read, Write and Seek traits for ease of use.
-
-pub use super::{Read, Seek, SeekFrom, Write};
