+++ src/object_safe.rs
@@ -1,40 +1,20 @@
 //! Object-safe traits for [`File`][], [`Filesystem`][] and [`Storage`][].
 
 use generic_array::typenum::Unsigned as _;
-use heapless::Vec;
 
 use crate::{
     driver::Storage,
-    fs::{Attribute, DirEntry, File, FileOpenFlags, Filesystem, Metadata},
-    io::{Error, OpenSeekFrom, Read, Result, Seek, Write},
+    fs::{Attribute, File, FileOpenFlags, Filesystem, Metadata},
+    io::{Error, OpenSeekFrom, Result},
     path::Path,
 };
 
+pub use littlefs2_core::{DirEntriesCallback, DynFile, DynFilesystem, FileCallback, Predicate};
+
 // Make sure that the traits actually are object safe.
-const _: Option<&dyn DynFile> = None;
-const _: Option<&dyn DynFilesystem> = None;
 const _: Option<&dyn DynStorage> = None;
 
-pub type DirEntriesCallback<'a, R = ()> =
-    &'a mut dyn FnMut(&mut dyn Iterator<Item = Result<DirEntry>>) -> Result<R>;
-pub type FileCallback<'a, R = ()> = &'a mut dyn FnMut(&dyn DynFile) -> Result<R>;
 pub type FilesystemCallback<'a, R = ()> = &'a mut dyn FnMut(&dyn DynFilesystem) -> Result<R>;
-pub type Predicate<'a> = &'a dyn Fn(&DirEntry) -> bool;
-
-/// Object-safe trait for [`File`][].
-///
-/// The methods for opening files cannot be implemented in this trait.  Use these methods instead:
-/// - [`DynFilesystem::create_file_and_then`](trait.DynFilesystem.html#method.create_file_and_then)
-/// - [`DynFilesystem::open_file_and_then`](trait.DynFilesystem.html#method.open_file_and_then)
-/// - [`DynFilesystem::open_file_with_options_and_then`](trait.DynFilesystem.html#method.open_file_with_options_and_then)
-///
-/// All other methods are mirrored directly.  See the documentation for [`File`][] for more information.
-pub trait DynFile: Read + Seek + Write {
-    fn sync(&self) -> Result<()>;
-    fn len(&self) -> Result<usize>;
-    fn is_empty(&self) -> Result<bool>;
-    fn set_len(&self, size: usize) -> Result<()>;
-}
 
 impl<S: Storage> DynFile for File<'_, '_, S> {
     fn sync(&self) -> Result<()> {
@@ -52,68 +32,6 @@ impl<S: Storage> DynFile for File<'_, '_, S> {
     fn set_len(&self, size: usize) -> Result<()> {
         File::set_len(self, size)
     }
-}
-
-impl dyn DynFile + '_ {
-    pub fn read_to_end<const N: usize>(&self, buf: &mut Vec<u8, N>) -> Result<usize> {
-        let had = buf.len();
-        buf.resize_default(buf.capacity()).unwrap();
-        let read = self.read(&mut buf[had..])?;
-        buf.truncate(had + read);
-        Ok(read)
-    }
-}
-
-/// Object-safe trait for [`Filesystem`][].
-///
-/// The following methods are implemented in [`DynStorage`][] instead:
-/// - [`DynStorage::format`][]
-/// - [`DynStorage::is_mountable`][]
-/// - [`DynStorage::mount_and_then`](trait.DynStorage.html#method.mount_and_then)
-///
-/// The following methods cannot support generic return types in the callbacks:
-/// - [`DynFilesystem::create_file_and_then_unit`][]
-/// - [`DynFilesystem::open_file_and_then_unit`][]
-/// - [`DynFilesystem::open_file_with_flags_and_then_unit`][]
-/// - [`DynFilesystem::read_dir_and_then_unit`][]
-///
-/// Use these helper functions instead:
-/// - [`DynFilesystem::create_file_and_then`](#method.create_file_and_then)
-/// - [`DynFilesystem::open_file_and_then`](#method.open_file_and_then)
-/// - [`DynFilesystem::open_file_with_options_and_then`](#method.open_file_with_options_and_then)
-/// - [`DynFilesystem::read_dir_and_then`](#method.read_dir_and_then)
-///
-/// All other methods are mirrored directly.  See the documentation for [`Filesystem`][] for more information.
-pub trait DynFilesystem {
-    fn total_blocks(&self) -> usize;
-    fn total_space(&self) -> usize;
-    fn available_blocks(&self) -> Result<usize>;
-    fn available_space(&self) -> Result<usize>;
-    fn remove(&self, path: &Path) -> Result<()>;
-    fn remove_dir(&self, path: &Path) -> Result<()>;
-    #[cfg(feature = "dir-entry-path")]
-    fn remove_dir_all(&self, path: &Path) -> Result<()>;
-    #[cfg(feature = "dir-entry-path")]
-    fn remove_dir_all_where(&self, path: &Path, predicate: Predicate<'_>) -> Result<usize>;
-    fn rename(&self, from: &Path, to: &Path) -> Result<()>;
-    fn exists(&self, path: &Path) -> bool;
-    fn metadata(&self, path: &Path) -> Result<Metadata>;
-    fn create_file_and_then_unit(&self, path: &Path, f: FileCallback<'_>) -> Result<()>;
-    fn open_file_and_then_unit(&self, path: &Path, f: FileCallback<'_>) -> Result<()>;
-    fn open_file_with_flags_and_then_unit(
-        &self,
-        flags: FileOpenFlags,
-        path: &Path,
-        f: FileCallback<'_>,
-    ) -> Result<()>;
-    fn attribute(&self, path: &Path, id: u8) -> Result<Option<Attribute>>;
-    fn remove_attribute(&self, path: &Path, id: u8) -> Result<()>;
-    fn set_attribute(&self, path: &Path, attribute: &Attribute) -> Result<()>;
-    fn read_dir_and_then_unit(&self, path: &Path, f: DirEntriesCallback<'_>) -> Result<()>;
-    fn create_dir(&self, path: &Path) -> Result<()>;
-    fn create_dir_all(&self, path: &Path) -> Result<()>;
-    fn write(&self, path: &Path, contents: &[u8]) -> Result<()>;
-    fn write_chunk(&self, path: &Path, contents: &[u8], pos: OpenSeekFrom) -> Result<()>;
 }
 
 impl<S: Storage> DynFilesystem for Filesystem<'_, S> {
@@ -218,73 +136,6 @@ impl<S: Storage> DynFilesystem for Filesystem<'_, S> {
 
     fn write_chunk(&self, path: &Path, contents: &[u8], pos: OpenSeekFrom) -> Result<()> {
         Filesystem::write_chunk(self, path, contents, pos)
-    }
-}
-
-impl dyn DynFilesystem + '_ {
-    pub fn read<const N: usize>(&self, path: &Path) -> Result<Vec<u8, N>> {
-        let mut contents = Vec::new();
-        self.open_file_and_then(path, &mut |file| {
-            file.read_to_end(&mut contents)?;
-            Ok(())
-        })?;
-        Ok(contents)
-    }
-
-    pub fn read_chunk<const N: usize>(
-        &self,
-        path: &Path,
-        pos: OpenSeekFrom,
-    ) -> Result<(Vec<u8, N>, usize)> {
-        let mut contents = Vec::new();
-        let file_len = self.open_file_and_then(path, &mut |file| {
-            file.seek(pos.into())?;
-            let read_n = file.read(&mut contents)?;
-            contents.truncate(read_n);
-            file.len()
-        })?;
-        Ok((contents, file_len))
-    }
-
-    pub fn create_file_and_then<R>(&self, path: &Path, f: FileCallback<'_, R>) -> Result<R> {
-        let mut result = Err(Error::IO);
-        self.create_file_and_then_unit(path, &mut |file| {
-            result = Ok(f(file)?);
-            Ok(())
-        })?;
-        result
-    }
-
-    pub fn open_file_and_then<R>(&self, path: &Path, f: FileCallback<'_, R>) -> Result<R> {
-        let mut result = Err(Error::IO);
-        self.open_file_and_then_unit(path, &mut |file| {
-            result = Ok(f(file)?);
-            Ok(())
-        })?;
-        result
-    }
-
-    pub fn open_file_with_flags_and_then<R>(
-        &self,
-        flags: FileOpenFlags,
-        path: &Path,
-        f: FileCallback<'_, R>,
-    ) -> Result<R> {
-        let mut result = Err(Error::IO);
-        self.open_file_with_flags_and_then_unit(flags, path, &mut |file| {
-            result = Ok(f(file)?);
-            Ok(())
-        })?;
-        result
-    }
-
-    pub fn read_dir_and_then<R>(&self, path: &Path, f: DirEntriesCallback<'_, R>) -> Result<R> {
-        let mut result = Err(Error::IO);
-        self.read_dir_and_then_unit(path, &mut |entries| {
-            result = Ok(f(entries)?);
-            Ok(())
-        })?;
-        result
     }
 }
